refactor: SnackReview 리팩토링 (#90)
