+++ server/src/main/java/seb/project/Codetech/snackreview/repository/CustomSnackReviewRepository.java
@@ -1,10 +1,12 @@
 package seb.project.Codetech.snackreview.repository;
 
+import java.util.List;
+
 import seb.project.Codetech.snackreview.dto.SnackReviewResponseDto;
 import seb.project.Codetech.snackreview.dto.SnackReviewServiceDto;
 
 public interface CustomSnackReviewRepository {
-	SnackReviewResponseDto.Slice searchSortedSliceByProductId(SnackReviewServiceDto.Search cond);
+	List<SnackReviewResponseDto.Card> searchSortedCardsByProductId(SnackReviewServiceDto.Search cond);
 
 	SnackReviewResponseDto.Info searchInfoGroupByProductId(Long productId);
 }
